chore: move independant functions from keeper.go and set emitTypedEventWithLog as private
